index.js
This is the file which handles all my intents. My helloWorld intent does the main job but because I used the hello World template, the name was left like that. I have lots to update within the skill. I am also a beginner in using Node.js libraries so my code may not be so efficient. I am using the Cheerio library and the request promise library too. This is a work in progress but do have a look and suggest any changes.